+++ packages/react-router/src/components/Link/Link.tsx
@@ -4,9 +4,10 @@ import {Link as ReactRouterLink} from 'react-router-dom';
 export interface Props {
   external?: boolean;
   url: string;
+  children?: React.ReactNode;
 }
 
-export default function Link({url, external, ...rest}: Props) {
+export default function Link({url, external, children, ...rest}: Props) {
   let target: string | undefined;
   let rel: string | undefined;
   if (external) {
@@ -14,5 +15,9 @@ export default function Link({url, external, ...rest}: Props) {
     rel = 'noopener noreferrer';
   }
 
-  return <ReactRouterLink target={target} to={url} rel={rel} {...rest} />;
+  return (
+    <ReactRouterLink target={target} to={url} rel={rel} {...rest}>
+      {children}
+    </ReactRouterLink>
+  );
 }